model_integration/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
Django==4.2.15
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
replicate==0.31.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
